requirements.txt
fastapi
uvicorn
sentence-transformers
keybert
scikit-learn
spacy
pydantic
python-dotenv
httpx
requests
google-api-python-client
beautifulsoup4
selenium
